requirements.txt
flask
flask-mobility==1.1.0
torch==1.13.1
datasets==2.12.0
transformers==4.28.1
sentencepiece==0.1.99
protobuf==3.20.1
urllib3==1.24
gunicorn==20.1.0
pandas==2.0.2
pg8000==1.29.8
python-dotenv==0.15.0
psycopg2-binary==2.9.6
cloud-sql-python-connector==1.2.4
SQLAlchemy==2.0.11